+++ src/DIC/doNotAutowire.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Thesis\DIC;
+
+/**
+ * @api
+ */
+const doNotAutowire = new Mapping\DoNotAutowire();